+++ kernel/char/tty/tty_ioctl.c
@@ -184,7 +184,7 @@ int tty_ioctl_int(tty *t, devfs_file_handle *h, uptr request, void *argp)
 int tty_fcntl_int(tty *t, devfs_file_handle *h, int cmd, uptr arg)
 {
    if (cmd == F_GETFL)
-      return h->flags;
+      return (int)h->flags;
 
    if (cmd == F_SETFL) {
       /*
